Split into different modules
This update has been tested through all cases and works properly. Still need to create REDCap_Functions.py and work on Send_Alerts/Daily_Updates.

Also, built in an adjustable lag for archiving alert that can be found in Our_Queries.py/Get_not_elevated_sensors() which is called in GetSort_Spikes.py

#24 #19
